requirements.txt
asgiref==3.5.2
Django==4.0.5
sqlparse==0.4.2
django-model-utils==4.2.0
Pillow==9.1.1
djangorestframework==3.13.1
drf-yasg==1.20.0
django-filter==22.1
uvicorn==0.17.6
gunicorn==20.1.0
